requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
pinecone==7.3.0
langchain==0.3.12
langchain-community==0.3.12
langchain-openai==0.2.2
openai==1.43.0
python-dotenv==1.0.0
scikit-learn==1.3.2
faiss-cpu==1.12.0
pymongo==4.6.1
sentence-transformers==2.7.0
pillow==10.2.0